+++ src/gex/toolbox.py
@@ -1,4 +1,6 @@
 import click
+import click_log
+import logging
 
 from gex.commands.archive.archive import archive
 from gex.commands.file.file import file
@@ -6,7 +8,11 @@ from gex.commands.finder.finder import finder
 from gex.commands.postprocess.postprocess import postprocess
 from gex.commands.version import version
 
+logger = logging.getLogger('gextoolbox')
+click_log.basic_config(logger)
+
 @click.group()
+@click_log.simple_verbosity_option(logger)
 def cli():
     pass